Move request arrange test utility to base test module
